persistent_skiplist fix bugs

+++ nvm_mod/pmem_manager.cc
@@ -6,7 +6,7 @@ PmemManager::PmemManager(NVMOption* nvm_option, std::string filename) {
   pmem_path = nvm_option->pmem_path;
   pmem_file_name = filename;
   OpenNVMFile();
-  memory_usage_ = (size_t*)GetMemoryUsage();
+  memory_usage_ = (uint32_t*)GetMemoryUsage();
   alloc_bytes_remaining_ = write_buffer_size - *memory_usage_;
   alloc_ptr_ = GetDataStart() + *memory_usage_;
 }
